# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Forma-te API Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "moises-alberto@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/ebooks": {
            "get": {
                "tags": [
                    "Ebooks"
                ],
                "summary": "Get a paginated list of ebooks",
                "description": "Returns a paginated list of ebooks based on the provided query parameters.",
                "operationId": "getAllEbook",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Additional filter for searching ebooks",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of ebooks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Ebook"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/ebook/{ebookId}": {
            "get": {
                "tags": [
                    "Ebooks"
                ],
                "summary": "Get ebook by ID",
                "description": "Returns details of a single ebook based on the provided ebook ID",
                "operationId": "getEbookById",
                "parameters": [
                    {
                        "name": "ebookId",
                        "in": "path",
                        "description": "ID of ebook to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ebook"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ebook"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ebooks"
                ],
                "summary": "Update a ebook",
                "description": "Updates an existing ebook based on the data provided in the request.",
                "operationId": "updateEbook",
                "parameters": [
                    {
                        "name": "ebookId",
                        "in": "path",
                        "description": "ID of the ebook to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Ebook data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "ebook Name"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://example.com/images/ebook.jpg"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ebook updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Updated ebook Name"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "Updated https://example.com/images/ebook.jpg"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "Updated ABC123"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 99.99
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ebook not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ebooks"
                ],
                "summary": "Delete a ebook",
                "description": "Deletes an existing ebook based on the provided ID.",
                "operationId": "destroyEbook",
                "parameters": [
                    {
                        "name": "ebookId",
                        "in": "path",
                        "description": "ID of the ebook to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ebook deleted successfully"
                    },
                    "404": {
                        "description": "ebook not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ebook": {
            "post": {
                "tags": [
                    "Ebooks"
                ],
                "summary": "Create a new ebook",
                "description": "Creates a new ebook based on the data provided in the request.",
                "operationId": "createEbook",
                "requestBody": {
                    "description": "ebook data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "ebook Name"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://example.com/images/ebook.jpg"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ebook created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "category_id": {
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "url": {
                                            "type": "string"
                                        },
                                        "image": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "category_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The category_id field is required."
                                                    }
                                                },
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The course name field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Ebook": {
                "title": "Ebook model",
                "description": "Ebook model",
                "required": [
                    "category_id",
                    "user_id",
                    "name"
                ],
                "properties": {
                    "category_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Ebook"
                }
            },
...
        }
    }
}